- removes the App folder / namespace - updates codacy badge
