+++ README.md
@@ -1,7 +1,7 @@
 <!--h-->
 # Control Panel Common
 
-[![Codacy Badge](https://api.codacy.com/project/badge/Grade/3200ceba8aea4a31ac1dfe826328bcb1)](https://www.codacy.com/app/laravel-enso/control-panel-common?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=laravel-enso/control-panel-common&amp;utm_campaign=Badge_Grade)
+[![Codacy Badge](https://app.codacy.com/project/badge/Grade/2e4f1f8a51c34019bddceaeed044f6db)](https://www.codacy.com/gh/laravel-enso/control-panel-common?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=laravel-enso/control-panel-common&amp;utm_campaign=Badge_Grade)
 [![StyleCI](https://github.styleci.io/repos/88979520/shield?branch=master)](https://github.styleci.io/repos/88979520)
 [![License](https://poser.pugx.org/laravel-enso/control-panel-common/license)](https://https://packagist.org/packages/laravel-enso/control-panel-common)
 [![Total Downloads](https://poser.pugx.org/laravel-enso/control-panel-common/downloads)](https://packagist.org/packages/laravel-enso/control-panel-common)
