+++ src/Http/Resources/Group.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace LaravelEnso\ControlPanelCommon\App\Http\Resources;
+namespace LaravelEnso\ControlPanelCommon\Http\Resources;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Str;
-use LaravelEnso\Helpers\App\Classes\Obj;
+use LaravelEnso\Helpers\Classes\Obj;
 
 class Group extends JsonResource
 {
